+++ gazebo/gui/model/JointMaker.cc
@@ -35,7 +35,9 @@
 #include "gazebo/gui/MouseEventHandler.hh"
 #include "gazebo/gui/GuiEvents.hh"
 #include "gazebo/gui/MainWindow.hh"
+#include "gazebo/gui/ModelAlign.hh"
 
+#include "gazebo/gui/model/JointCreationDialog.hh"
 #include "gazebo/gui/model/JointInspector.hh"
 #include "gazebo/gui/model/ModelEditorEvents.hh"
 #include "gazebo/gui/model/JointMaker.hh"
@@ -77,6 +79,8 @@ JointMaker::JointMaker()
   this->jointTypes[JOINT_BALL]      = "ball";
   this->jointTypes[JOINT_GEARBOX]   = "gearbox";
   this->jointTypes[JOINT_NONE]      = "none";
+
+  this->jointCreationDialog = NULL;
 
   this->connections.push_back(
       event::Events::ConnectPreRender(
@@ -132,6 +136,10 @@ JointMaker::~JointMaker()
     this->joints.clear();
   }
 
+  if (this->jointCreationDialog)
+    delete this->jointCreationDialog;
+
+  // Delete this last
   delete this->updateMutex;
 }
 
@@ -351,7 +359,11 @@ bool JointMaker::OnMousePress(const common::MouseEvent &_event)
 bool JointMaker::OnMouseRelease(const common::MouseEvent &_event)
 {
   rendering::UserCameraPtr camera = gui::get_active_camera();
-  if (this->jointType == JointMaker::JOINT_NONE)
+
+  // Not in the process of selecting joint links with mouse
+  // Handle joint selection
+  if (this->jointType == JointMaker::JOINT_NONE ||
+      (this->newJoint && this->newJoint->parent && this->newJoint->child))
   {
     rendering::VisualPtr vis = camera->GetVisual(_event.Pos());
     if (vis)
@@ -389,6 +401,7 @@ bool JointMaker::OnMouseRelease(const common::MouseEvent &_event)
       return false;
     }
   }
+  // Still selecting parent/child during new joint creation
   else
   {
     if (_event.Button() == common::MouseEvent::LEFT)
@@ -407,12 +420,24 @@ bool JointMaker::OnMouseRelease(const common::MouseEvent &_event)
         {
           if (!this->SetParentLink(this->hoverVis))
             return false;
+
+          if (this->jointCreationDialog)
+          {
+            this->jointCreationDialog->SetParent(
+                this->newJoint->parent->GetName());
+          }
         }
         // Pressed child link
         else if (this->newJoint && this->newJoint->parent != this->hoverVis)
         {
           if (!this->SetChildLink(this->hoverVis))
             return false;
+
+          if (this->jointCreationDialog)
+          {
+            this->jointCreationDialog->SetChild(
+                this->newJoint->child->GetName());
+          }
         }
 
         if (this->hoverVis)
@@ -523,7 +548,18 @@ void JointMaker::AddJoint(const std::string &_type)
 void JointMaker::AddJoint(JointMaker::JointType _type)
 {
   this->jointType = _type;
-  if (_type == JointMaker::JOINT_NONE)
+  // Start joint creation
+  if (_type != JointMaker::JOINT_NONE)
+  {
+    if (!this->jointCreationDialog)
+    {
+      auto mainWindow = gui::get_main_window();
+      this->jointCreationDialog = new JointCreationDialog(this, mainWindow);
+    }
+    this->jointCreationDialog->Open(_type);
+  }
+  // End joint creation
+  else
   {
     // signal the end of a joint action.
     emit JointAdded();
@@ -533,30 +569,42 @@ void JointMaker::AddJoint(JointMaker::JointType _type)
 /////////////////////////////////////////////////
 void JointMaker::Stop()
 {
-  if (this->jointType != JointMaker::JOINT_NONE)
+  // Reset links
+  if (this->newJoint)
   {
-    if (this->newJoint)
+    if (this->newJoint->parent)
     {
-      this->newJoint->visual->DeleteDynamicLine(this->newJoint->line);
-      rendering::ScenePtr scene = this->newJoint->visual->GetScene();
-      scene->RemoveVisual(this->newJoint->visual);
-      this->newJoint->visual.reset();
-      delete this->newJoint;
-      this->newJoint = NULL;
+      this->newJoint->parent->SetWorldPose(this->parentLinkOriginalPose);
+      this->SetVisualMoved(this->newJoint->parent, false);
     }
-    this->AddJoint(JointMaker::JOINT_NONE);
-    if (this->hoverVis)
-      this->hoverVis->SetEmissive(common::Color(0, 0, 0));
-    this->hoverVis.reset();
+    if (this->newJoint->child)
+    {
+      this->newJoint->child->SetWorldPose(this->childLinkOriginalPose);
+      this->SetVisualMoved(this->newJoint->child, false);
+    }
   }
+
+  this->RemoveJoint("");
+
+  this->AddJoint(JointMaker::JOINT_NONE);
+  if (this->hoverVis)
+    this->hoverVis->SetEmissive(common::Color(0, 0, 0));
+  this->hoverVis.reset();
+
+  if (this->jointCreationDialog && this->jointCreationDialog->isVisible())
+    this->jointCreationDialog->reject();
 }
 
 /////////////////////////////////////////////////
 bool JointMaker::OnMouseMove(const common::MouseEvent &_event)
 {
   // Only handle mouse move events during joint creation
-  if (this->jointType == JointMaker::JOINT_NONE)
+  if (this->jointType == JointMaker::JOINT_NONE ||
+      (this->newJoint && this->newJoint->child))
+  {
     return false;
+  }
+  QApplication::setOverrideCursor(Qt::CrossCursor);
 
   // Get the active camera and scene.
   rendering::UserCameraPtr camera = gui::get_active_camera();
@@ -693,10 +741,10 @@ void JointMaker::OnDelete()
 }
 
 /////////////////////////////////////////////////
-void JointMaker::CreateHotSpot(JointData *_joint)
+std::string JointMaker::CreateHotSpot(JointData *_joint)
 {
   if (!_joint)
-    return;
+    return "";
 
   rendering::UserCameraPtr camera = gui::get_active_camera();
 
@@ -753,11 +801,7 @@ void JointMaker::CreateHotSpot(JointData *_joint)
 
   _joint->dirty = true;
 
-  std::string parentName = _joint->parent->GetName();
-  std::string childName = _joint->child->GetName();
-
-  gui::model::Events::jointInserted(jointId, _joint->name,
-      jointTypes[_joint->type], parentName, childName);
+  return jointId;
 }
 
 /////////////////////////////////////////////////
@@ -781,12 +825,45 @@ void JointMaker::Update()
            joint->childPose = joint->child->GetWorldPose();
            joint->childScale = joint->child->GetScale();
            poseUpdate = true;
+
+           // Highlight links connected to joint being created if they have
+           // been moved to another position
+           if (joint == this->newJoint)
+           {
+             // Parent
+             this->SetVisualMoved(joint->parent,
+                 joint->parent->GetWorldPose().Ign() !=
+                 this->parentLinkOriginalPose);
+
+             // Child
+             this->SetVisualMoved(joint->child,
+                 joint->child->GetWorldPose().Ign() !=
+                 this->childLinkOriginalPose);
+           }
          }
 
         // Create / update joint visual
         if (joint->dirty || poseUpdate)
         {
           joint->Update();
+
+          if (joint == this->newJoint && this->newJoint->parent &&
+              this->newJoint->child && this->jointCreationDialog &&
+              this->jointCreationDialog->isVisible())
+          {
+            // Get poses as homogeneous transforms
+            ignition::math::Matrix4d parentWorld(
+                this->newJoint->parent->GetWorldPose().Ign());
+            ignition::math::Matrix4d childWorld(
+                this->newJoint->child->GetWorldPose().Ign());
+
+            // w_T_c = w_T_p * p_T_c
+            // w_T_p^-1 * w_T_c = p_T_c
+            ignition::math::Matrix4d childParent = parentWorld.Inverse() *
+                childWorld;
+
+            this->jointCreationDialog->UpdateRelativePose(childParent.Pose());
+          }
         }
       }
     }
@@ -1083,7 +1160,7 @@ void JointData::Update()
       this->jointVisual->UpdateFromMsg(this->jointMsg);
     }
     // Create joint visual
-    else
+    else if (this->child)
     {
       std::string childName = this->child->GetName();
       std::string jointVisName = childName;
@@ -1229,8 +1306,12 @@ void JointData::UpdateMsg()
       }
       else
       {
-        msgs::Set(axisMsg->mutable_xyz(),
-            JointMaker::unitVectors[i%JointMaker::unitVectors.size()]);
+        if (this->axes.size() < i+1)
+        {
+          this->axes.push_back(
+              JointMaker::unitVectors[i%JointMaker::unitVectors.size()]);
+        }
+        msgs::Set(axisMsg->mutable_xyz(), this->axes[i]);
       }
       axisMsg->set_use_parent_model_frame(false);
       axisMsg->set_limit_lower(-GZ_DBL_MAX);
@@ -1403,7 +1484,15 @@ void JointMaker::CreateJointFromSDF(sdf::ElementPtr _jointElem,
   joint->line = jointLine;
   joint->dirty = true;
 
-  this->CreateHotSpot(joint);
+  auto jointId = this->CreateHotSpot(joint);
+
+  // Notify other widgets
+  if (!jointId.empty())
+  {
+    gui::model::Events::jointInserted(jointId, joint->name,
+        jointTypes[joint->type], joint->parent->GetName(),
+        joint->child->GetName());
+  }
 }
 
 /////////////////////////////////////////////////
@@ -1460,17 +1549,30 @@ bool JointMaker::SetParentLink(rendering::VisualPtr _parentLink)
 
   boost::recursive_mutex::scoped_lock lock(*this->updateMutex);
 
+  // Choosing parent for the first time
   if (!this->newJoint)
   {
     // Create new line connecting parent to mouse
     this->newJoint = this->CreateJointLine("JOINT_LINE", _parentLink);
   }
+  // Update parent of joint being created
+  else if (this->newJoint->parent)
+  {
+    // Reset previous parent
+    this->newJoint->parent->SetWorldPose(this->parentLinkOriginalPose);
+    this->SetVisualMoved(this->newJoint->parent, false);
+
+    this->newJoint->parent = _parentLink;
+    this->newJoint->dirty = true;
+  }
   else
   {
-    gzerr << "Currently it's not possible to change the parent of the joint "
-        << "being created." << std::endl;
+    gzerr << "There's a joint being created but the parent visual hasn't been "
+        << "defined. This should never happen." << std::endl;
     return false;
   }
+
+  this->parentLinkOriginalPose = _parentLink->GetWorldPose().Ign();
   return true;
 }
 
@@ -1491,6 +1593,7 @@ bool JointMaker::SetChildLink(rendering::VisualPtr _childLink)
 
   boost::recursive_mutex::scoped_lock lock(*this->updateMutex);
 
+  // Choosing child for the first time
   if (!this->newJoint->child)
   {
     rendering::VisualPtr parentVis = this->newJoint->parent;
@@ -1500,23 +1603,224 @@ bool JointMaker::SetChildLink(rendering::VisualPtr _childLink)
 
     // Create new joint with parent and child
     auto joint = this->CreateJoint(parentVis, _childLink);
+    this->newJoint = joint;
 
-    this->CreateHotSpot(joint);
-
-    // Finish joint creation
-    this->newJoint = NULL;
-    gui::model::Events::modelChanged();
+    // Create hotspot visual
+    this->CreateHotSpot(this->newJoint);
   }
   // Update child
   else
   {
-    gzerr << "Currently it's not possible to change the child of the joint "
-        << "being created." << std::endl;
-    return false;
+    // Reset previous child
+    this->newJoint->child->SetWorldPose(this->childLinkOriginalPose);
+    this->SetVisualMoved(this->newJoint->child, false);
+
+    this->newJoint->child = _childLink;
+    this->newJoint->dirty = true;
+    _childLink->AttachVisual(this->newJoint->jointVisual);
   }
+
+  this->childLinkOriginalPose = _childLink->GetWorldPose().Ign();
 
   // Change state to not creating joint
   gui::Events::manipMode("select");
   this->jointType = JointMaker::JOINT_NONE;
   return true;
+}
+
+/////////////////////////////////////////////////
+void JointMaker::OnType(const int _typeInt)
+{
+  this->jointType = static_cast<JointMaker::JointType>(_typeInt);
+
+  if (this->newJoint && this->jointType != JOINT_NONE)
+  {
+    this->newJoint->type = this->jointType;
+    this->newJoint->dirty = true;
+  }
+}
+
+/////////////////////////////////////////////////
+void JointMaker::SetAxis(const QString &_axis,
+      const ignition::math::Vector3d &_value)
+{
+  if (this->newJoint && this->newJoint->jointMsg)
+  {
+    if (_axis == "axis1" && this->newJoint->jointMsg->has_axis1())
+    {
+      msgs::Set(this->newJoint->jointMsg->mutable_axis1()->mutable_xyz(),
+        _value);
+      this->newJoint->axes[0] = _value;
+    }
+    else if (_axis == "axis2" && this->newJoint->jointMsg->has_axis2())
+    {
+      msgs::Set(this->newJoint->jointMsg->mutable_axis2()->mutable_xyz(),
+        _value);
+      this->newJoint->axes[1] = _value;
+    }
+    this->newJoint->dirty = true;
+  }
+}
+
+/////////////////////////////////////////////////
+void JointMaker::SetJointPose(const ignition::math::Pose3d &_pose)
+{
+  if (this->newJoint && this->newJoint->jointMsg)
+  {
+    msgs::Set(this->newJoint->jointMsg->mutable_pose(), _pose);
+    this->newJoint->dirty = true;
+  }
+}
+
+/////////////////////////////////////////////////
+void JointMaker::SetParentLink(const std::string &_name)
+{
+  auto vis = this->LinkVisualFromName(_name);
+
+  if (vis)
+    this->SetParentLink(vis);
+}
+
+/////////////////////////////////////////////////
+void JointMaker::SetChildLink(const std::string &_name)
+{
+  auto vis = this->LinkVisualFromName(_name);
+
+  if (vis)
+    this->SetChildLink(vis);
+}
+
+/////////////////////////////////////////////////
+rendering::VisualPtr JointMaker::LinkVisualFromName(const std::string &_name)
+{
+  // Get scoped name
+  std::string scopedName;
+  for (auto link : this->linkList)
+  {
+    if (link.second == _name || link.first == _name)
+    {
+      scopedName = link.first;
+      break;
+    }
+  }
+
+  if (scopedName.empty())
+    return NULL;
+
+  // Get visual
+  rendering::ScenePtr scene = rendering::get_scene();
+
+  if (!scene)
+    return NULL;
+
+  return scene->GetVisual(scopedName);
+}
+
+/////////////////////////////////////////////////
+void JointMaker::SetLinksRelativePose(const ignition::math::Pose3d &_pose,
+    const bool _reset)
+{
+  if (!this->newJoint || !this->newJoint->parent || !this->newJoint->child)
+  {
+    gzerr << "Can't set relative pose without new joint's parent and child "
+        << "links." << std::endl;
+    return;
+  }
+
+  ignition::math::Pose3d newChildPose;
+
+  if (_reset)
+  {
+    newChildPose = this->childLinkOriginalPose;
+    this->newJoint->parent->SetWorldPose(this->parentLinkOriginalPose);
+  }
+  else
+  {
+    // Get poses as homogeneous transforms
+    ignition::math::Matrix4d parent_world(
+        this->newJoint->parent->GetWorldPose().Ign());
+    ignition::math::Matrix4d child_parent(_pose);
+
+    // w_T_c = w_T_p * p_T_c
+    ignition::math::Matrix4d child_world =
+        parent_world * child_parent;
+
+    newChildPose = child_world.Pose();
+  }
+
+  this->newJoint->child->SetWorldPose(newChildPose);
+}
+
+/////////////////////////////////////////////////
+void JointMaker::AlignLinks(const bool _childToParent,
+    const std::string &_axis, const std::string &_config)
+{
+  if (!this->newJoint || !this->newJoint->parent || !this->newJoint->child)
+  {
+    gzerr << "Couldn't find new joint's parent and child links to be aligned."
+        << std::endl;
+    return;
+  }
+
+  std::vector<rendering::VisualPtr> links;
+  links.push_back(this->newJoint->parent);
+  links.push_back(this->newJoint->child);
+
+  std::string target = _childToParent ? "first" : "last";
+
+  ModelAlign::Instance()->AlignVisuals(links, _axis, _config,
+      target, true);
+}
+
+/////////////////////////////////////////////////
+void JointMaker::SetVisualMoved(rendering::VisualPtr _vis, const bool _moved)
+{
+  if (_vis->GetChildCount() != 0)
+  {
+    for (unsigned int j = 0; j < _vis->GetChildCount(); ++j)
+    {
+      this->SetVisualMoved(_vis->GetChild(j), _moved);
+    }
+  }
+  else
+  {
+    if (_moved)
+    {
+      _vis->SetEmissive(common::Color(0, 0, 1, 1));
+    }
+    else
+    {
+      _vis->SetEmissive(common::Color(0, 0, 0, 1));
+    }
+  }
+}
+
+/////////////////////////////////////////////////
+void JointMaker::FinalizeCreation()
+{
+  gui::model::Events::modelChanged();
+  this->jointType = JointMaker::JOINT_NONE;
+
+  // Notify schematic view and palette list
+  if (this->newJoint && this->newJoint->hotspot &&
+      this->newJoint->child && this->newJoint->parent)
+  {
+    gui::model::Events::jointInserted(
+        this->newJoint->hotspot->GetName(),
+        this->newJoint->name,
+        this->jointTypes[this->newJoint->type],
+        this->newJoint->parent->GetName(),
+        this->newJoint->child->GetName());
+
+
+    // Reset visuals
+    this->SetVisualMoved(this->newJoint->parent, false);
+    this->SetVisualMoved(this->newJoint->child, false);
+  }
+  this->newJoint = NULL;
+
+  // Notify ModelEditor to uncheck tool button
+  this->JointAdded();
+
+  this->Stop();
 }